+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 0.4.0 (under development)
 -------------------------
+* ftl2elm --include option
+
 
 0.3.0 (2019-03-06)
 ------------------